services/order-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pytest
psycopg2-binary

